+++ src/extension.ts
@@ -17,6 +17,7 @@ import { Parser } from './parser';
 import { Rules } from './rules';
 
 // Language specific code parsers
+import { C } from './languages/c';
 import { Java } from './languages/java';
 import { JavaScript } from './languages/javascript';
 import { PHP } from './languages/php';
@@ -28,11 +29,13 @@ export function activate(context: ExtensionContext) {
   // Scheme as follows:
   //   language ID: class name
   const langList = {
+    c: C,
     java: Java,
     javascript: JavaScript,
     php: PHP,
     scss: Scss,
     typescript: TypeScript,
+    vue: TypeScript,
   };
   // Register each language
   for (const language in langList) {