+++ src/languages/java.ts
@@ -4,7 +4,8 @@
 
 'use strict';
 
-import { Param, Parser, Tokens } from '../parser';
+import { Parser } from '../parser';
+import { Param, Tokens } from '../tokens';
 
 export class Java extends Parser {
   /**
@@ -38,11 +39,11 @@ export class Java extends Parser {
    *
    * @return  {Tokens}          Tokens retrieved from Lexer output
    */
-  public tokenize(code: string, next: string = '', tokens: Tokens = null): Tokens {
-    // Create empty token object if none is present
-    if (tokens === null) {
-      tokens = {name: '', type: '', params: [], return: { present: true }};
-    }
+  public tokenize(
+    code: string,
+    next: string = '',
+    tokens = new Tokens(),
+  ): Tokens {
     // Don't continue unless we have workable value
     if (code !== undefined) {
       // Get logical OR expression for determining variables type. This is