+++ test/languages/c.test.ts
@@ -1,15 +1,14 @@
 /**
- * Tests specific to parsing the Java language
+ * Tests specific to parsing the C language
  */
 
 import * as assert from 'assert';
-import { Java } from '../../src/languages/java';
+import { C } from '../../src/languages/c';
 
-// Get parser instance
-const parser = new Java();
+const parser = new C();
 
-suite('Java', () => {
-  suite('tokenize', () => {
+suite('C', () => {
+  suite('Tokenize', () => {
     test('should parse variable', () => {
       const token = parser.tokenize('int foo = 5;');
 
@@ -21,27 +20,27 @@ suite('Java', () => {
     });
 
     test('should parse undefined variable', () => {
-      const token = parser.tokenize('boolean foo;');
+      const token = parser.tokenize('int foo;');
 
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'variable');
-      assert.equal(token.varType, 'boolean');
+      assert.equal(token.varType, 'int');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, false);
     });
 
     test('should parse function', () => {
-      const token = parser.tokenize('public void foo() {');
+      const token = parser.tokenize('char foo() {');
 
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 0);
-      assert.equal(token.return.type, 'void');
+      assert.equal(token.return.type, 'char');
       assert.equal(token.return.present, true);
     });
 
     test('should parse function with arguments', () => {
-      const token = parser.tokenize('public void foo(int arg1, int arg2) {');
+      const token = parser.tokenize('int foo(char arg1, char arg2) {');
 
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
@@ -50,36 +49,36 @@ suite('Java', () => {
         if (token.params[i]) {
           assert.equal(token.params[i].name, `arg${parseInt(i) + 1}`);
           assert.equal(token.params[i].val, '');
-          assert.equal(token.params[i].type, 'int');
+          assert.equal(token.params[i].type, 'char');
         }
       }
       assert.equal(token.return.present, true);
     });
 
-    test('should parser function with multiple modifiers', () => {
-      const token = parser.tokenize('public static void foo() {');
+    test('should parse function with multiple modifiers', () => {
+      const token = parser.tokenize('complex static int foo() {');
 
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 0);
-      assert.equal(token.return.type, 'void');
+      assert.equal(token.return.type, 'int');
       assert.equal(token.return.present, true);
     });
 
-    test('should parse class', () => {
-      const token = parser.tokenize('class Bar {');
+    test('should parse struct', () => {
+      const token = parser.tokenize('struct foo {');
 
-      assert.equal(token.name, 'Bar');
-      assert.equal(token.type, 'class');
+      assert.equal(token.name, 'struct');
+      assert.equal(token.type, 'struct');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, false);
     });
 
-    test('should parse abstract class', () => {
-      const token = parser.tokenize('abstract class Bar {');
+    test('should parse typedef struct', () => {
+      const token = parser.tokenize('typedef struct {');
 
-      assert.equal(token.name, 'Bar');
-      assert.equal(token.type, 'class');
+      assert.equal(token.name, 'struct');
+      assert.equal(token.type, 'struct');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, false);
     });