+++ test/languages/javascript.test.ts
@@ -2,74 +2,85 @@
  * Tests specific to parsing the JavaScript language
  */
 
-import * as assert    from 'assert';
+import * as assert from 'assert';
 import { JavaScript } from '../../src/languages/javascript';
 
 // Get parser instance
-let parser = new JavaScript();
+const parser = new JavaScript();
 
-suite('JavaScript', function () {
-  suite('tokenize', function () {
-    test('should parse variable', function () {
-      let token = parser.tokenize('let foo = 5');
+suite('JavaScript', () => {
+  suite('tokenize', () => {
+    test('should parse variable', () => {
+      const token = parser.tokenize('let foo = 5');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'variable');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, false);
     });
 
-    test('should parse undefined variable', function () {
-      let token = parser.tokenize('let foo');
+    test('should parse undefined variable', () => {
+      const token = parser.tokenize('let foo');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'variable');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, false);
     });
 
-    test('should parse function', function () {
-      let token = parser.tokenize('function foo() {');
+    test('should parse function', () => {
+      const token = parser.tokenize('function foo() {');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, true);
     });
 
-    test('should parse function with arguments', function () {
-      let token = parser.tokenize('function foo(arg1, arg2) {');
+    test('should parse function with arguments', () => {
+      const token = parser.tokenize('function foo(arg1, arg2) {');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 2);
-      for (let i in token.params) {
-        assert.equal(token.params[i].name, `arg${parseInt(i) + 1}`);
-        assert.equal(token.params[i].val, '');
-        assert.equal(token.params[i].type, undefined);
+      for (const i in token.params) {
+        if (token.params[i]) {
+          assert.equal(token.params[i].name, `arg${parseInt(i) + 1}`);
+          assert.equal(token.params[i].val, '');
+          assert.equal(token.params[i].type, undefined);
+        }
       }
       assert.equal(token.return.present, true);
     });
 
-    test('should parse class', function () {
-      let token = parser.tokenize('class Bar {');
+    test('should parse class', () => {
+      const token = parser.tokenize('class Bar {');
+
       assert.equal(token.name, 'Bar');
       assert.equal(token.type, 'class');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, false);
     });
 
-    test('should parse prototype function', function () {
-      let token = parser.tokenize('Bar.prototype.foo = function(arg1, arg2) {');
+    test('should parse prototype function', () => {
+      const token = parser.tokenize('Bar.prototype.foo = function(arg1, arg2) {');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 2);
-      for (let i in token.params) {
-        assert.equal(token.params[i].name, `arg${parseInt(i) + 1}`);
-        assert.equal(token.params[i].val, '');
-        assert.equal(token.params[i].type, undefined);
+      for (const i in token.params) {
+        if (token.params[i]) {
+          assert.equal(token.params[i].name, `arg${parseInt(i) + 1}`);
+          assert.equal(token.params[i].val, '');
+          assert.equal(token.params[i].type, undefined);
+        }
       }
       assert.equal(token.return.present, true);
     });
 
-    test('should parse expression assigned to object property', function () {
-      let token = parser.tokenize('Fizz.buzz.foo = function (bar) {');
+    test('should parse expression assigned to object property', () => {
+      const token = parser.tokenize('Fizz.buzz.foo = function (bar) {');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 1);