+++ test/languages/php.test.ts
@@ -2,45 +2,50 @@
  * Tests specific to parsing the PHP language
  */
 
-import * as assert    from 'assert';
-import { PHP }        from '../../src/languages/php';
+import * as assert from 'assert';
+import { PHP } from '../../src/languages/php';
 
 // Get parser instance
-let parser = new PHP();
+const parser = new PHP();
 
-suite('PHP', function () {
-  suite('tokenize', function () {
-    test('should parse variable', function () {
-      let token = parser.tokenize('$foo = 5');
+suite('PHP', () => {
+  suite('tokenize', () => {
+    test('should parse variable', () => {
+      const token = parser.tokenize('$foo = 5');
+
       assert.equal(token.name, '$foo');
       assert.equal(token.type, 'variable');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, false);
     });
 
-    test('should parse function', function () {
-      let token = parser.tokenize('function foo() {');
+    test('should parse function', () => {
+      const token = parser.tokenize('function foo() {');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, true);
     });
 
-    test('should parse function with arguments', function () {
-      let token = parser.tokenize('function foo($arg1, $arg2) {');
+    test('should parse function with arguments', () => {
+      const token = parser.tokenize('function foo($arg1, $arg2) {');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 2);
-      for (let i in token.params) {
-        assert.equal(token.params[i].name, `$arg${parseInt(i) + 1}`);
-        assert.equal(token.params[i].val, '');
-        assert.equal(token.params[i].type, undefined);
+      for (const i in token.params) {
+        if (token.params[i]) {
+          assert.equal(token.params[i].name, `$arg${parseInt(i) + 1}`);
+          assert.equal(token.params[i].val, '');
+          assert.equal(token.params[i].type, undefined);
+        }
       }
       assert.equal(token.return.present, true);
     });
 
-    test('should parse function with arguments passed by reference', function () {
-      let token = parser.tokenize('function foo(int &$arg): boolean {');
+    test('should parse function with arguments passed by reference', () => {
+      const token = parser.tokenize('function foo(int &$arg): boolean {');
 
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
@@ -54,48 +59,56 @@ suite('PHP', function () {
       assert.equal(token.return.type, 'boolean');
     });
 
-    test('should parse defined argument type', function () {
-      let token = parser.tokenize('function foo(int $bar = 0) {');
+    test('should parse defined argument type', () => {
+      const token = parser.tokenize('function foo(int $bar = 0) {');
+
       assert.equal(token.params[0].name, '$bar');
       assert.equal(token.params[0].type, 'int');
       assert.equal(token.params[0].val, '0');
     });
 
-    test('should parse defined return type', function () {
-      let token = parser.tokenize('function foo(): boolean {');
+    test('should parse defined return type', () => {
+      const token = parser.tokenize('function foo(): boolean {');
+
       assert.equal(token.return.present, true);
       assert.equal(token.return.type, 'boolean');
     });
 
-    test('should parse class name as return type', function () {
-      let token = parser.tokenize('function foo(): TestClass {');
+    test('should parse class name as return type', () => {
+      const token = parser.tokenize('function foo(): TestClass {');
+
       assert.equal(token.return.present, true);
       assert.equal(token.return.type, 'TestClass');
     });
 
-    test('should parse class', function () {
-      let token = parser.tokenize('class Bar {');
+    test('should parse class', () => {
+      const token = parser.tokenize('class Bar {');
+
       assert.equal(token.name, 'Bar');
       assert.equal(token.type, 'class');
       assert.equal(token.params.length, 0);
       assert.equal(token.return.present, false);
     });
 
-    test('should parse class method', function () {
-      let token = parser.tokenize('public function foo($arg1, $arg2) {');
+    test('should parse class method', () => {
+      const token = parser.tokenize('public function foo($arg1, $arg2) {');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 2);
-      for (let i in token.params) {
-        assert.equal(token.params[i].name, `$arg${parseInt(i) + 1}`);
-        assert.equal(token.params[i].val, '');
-        assert.equal(token.params[i].type, undefined);
+      for (const i in token.params) {
+        if (token.params[i]) {
+          assert.equal(token.params[i].name, `$arg${parseInt(i) + 1}`);
+          assert.equal(token.params[i].val, '');
+          assert.equal(token.params[i].type, undefined);
+        }
       }
       assert.equal(token.return.present, true);
     });
 
-    test('should parse class method argument type', function () {
-      let token = parser.tokenize('public function foo(string $arg1, stdClass $arg2) {');
+    test('should parse class method argument type', () => {
+      const token = parser.tokenize('public function foo(string $arg1, stdClass $arg2) {');
+
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 2);
@@ -110,22 +123,24 @@ suite('PHP', function () {
       assert.equal(token.return.present, true);
     });
 
-    test('should parse class method with return type', function () {
-      let token = parser.tokenize('public function foo($arg1, $arg2): boolean {');
+    test('should parse class method with return type', () => {
+      const token = parser.tokenize('public function foo($arg1, $arg2): boolean {');
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');
       assert.equal(token.params.length, 2);
-      for (let i in token.params) {
-        assert.equal(token.params[i].name, `$arg${parseInt(i) + 1}`);
-        assert.equal(token.params[i].val, '');
-        assert.equal(token.params[i].type, undefined);
+      for (const i in token.params) {
+        if (token.params[i]) {
+          assert.equal(token.params[i].name, `$arg${parseInt(i) + 1}`);
+          assert.equal(token.params[i].val, '');
+          assert.equal(token.params[i].type, undefined);
+        }
       }
       assert.equal(token.return.present, true);
       assert.equal(token.return.type, 'boolean');
     });
 
-    test('should parse class method with type defined arguments', function () {
-      let token = parser.tokenize('public function foo(int $arg): boolean {');
+    test('should parse class method with type defined arguments', () => {
+      const token = parser.tokenize('public function foo(int $arg): boolean {');
 
       assert.equal(token.name, 'foo');
       assert.equal(token.type, 'function');